adk_service/pyproject.toml
[project]
name = "lease-reviewer"
version = "0.1.0"
description = "Lease document reviewer using Google ADK + LiteLLM (Gemini 2.5 Flash)"
requires-python = ">=3.11"

dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.56.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lease_reviewer"]
